+++ backend/movie-ticket-booking-system/src/main/java/com/jts/movie/repositories/TheaterRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TheaterRepository extends JpaRepository<Theater, Integer> {
+public interface TheaterRepository extends JpaRepository<Theater, Long> {
 }
